bips instead of pips (#75)
* bips instead of pips

* workaround for fixing the computation

* snapshots

* update PriceMath.sol to handle bips

* gas tests

* make all functions private so they can be optimized by using uint256

* address comments

+++ contracts/interfaces/IUniswapV3Pair.sol
@@ -36,7 +36,7 @@ interface IUniswapV3Pair {
     function price1CumulativeLast() external view returns (uint256);
 
     // derived state
-    function getFee() external view returns (uint24 fee);
+    function getFee() external view returns (uint16 fee);
 
     function getVirtualSupply() external view returns (uint112 virtualSupply);
